Ok...fixed that mapping issue with the json req. Apparently Spring/Jaxson/(any json rest lib/framework) doesn't like it when you used fields that start with capital letters....
Ok...so now what do I need to work on ?
1. Figuring out what to do with the marathon.json file.
I feel that it for now should be included within the jar

2. Cleaning up code.  I will just be renaming things.  I am going to keep
the json and URI code seperate. Yeah...some may say that it is bad practice...
but later for them.  This is utility code, that isn't meant to be maintained by
a software engineer.  I want to make it as easy to deal with as possible.

I will leave it up to future maintainers to decide if they want to consoldate the "redundant" code
or not.

3. Need to test everything...fo course...